Move interceptor to utils package, make SynchronizedServerCall a static nested class
